Remove outdated userRepo tests
