+++ test/scripts-test.js
@@ -1,3 +1,12 @@
-// import { expect } from 'chai';
-// import '../src/scripts.js';
+const chai = require('chai')
+const spies = require('chai-spies');
+chai.use(spies); 
 
+describe('scripts', () => {
+  beforeEach(() => {
+    const windowRef = global.window; 
+    global.window = {document: {querySelector: () => {}}}
+    global.window = windowRef
+
+  })
+})
