+++ test/userrepo-test.js
@@ -43,16 +43,6 @@ describe('UserRepo', function() {
     expect(userRepo.data).to.deep.equal(users)
   });
 
-/* THE EXPECTATION HERE IS AN INSTANCEOF RESULT BECAUSE OF THE getRandomNumber
-METHOD IT UTILIZES. THE USER IT GRABS IS ALMOST ALWAYS A DIFFERENT ONE */
-  it.only('should be able to find the current user', function() {
-    userRepo.storeData(users)
-
-    expect(userRepo.currentUser).to.equal(null)
-    userRepo.findCurrentUser()
-    expect(userRepo.currentUser).to.be.an.instanceof(User)
-  });
-
   it.only('should be able to find a user by their ID', function() {
     userRepo.storeData(users)
     userRepo.findUserById(1)
@@ -60,14 +50,4 @@ METHOD IT UTILIZES. THE USER IT GRABS IS ALMOST ALWAYS A DIFFERENT ONE */
 
     expect(results.name).to.equal('Joshua Danger Sevy')
   });
-
-  it.only('should return a random number', function() {
-    userRepo.getRandomNumber()
-
-    expect(userRepo.getRandomNumber()).to.be.a('number')
-  });
-
-  // it.only('', function() {
-  //   expect().to.equal()
-  // })
 })
